Add SAveSettings method in DataSource

+++ QuranGardens/Settings.h
@@ -21,7 +21,8 @@ typedef NS_OPTIONS(NSUInteger, SorterType) {
 
 @interface Settings : NSObject
 
-@property (nonatomic) unsigned long fadeTime;
+/** Sura Cell fade time in seconds*/
+@property (nonatomic) double fadeTime;
 @property (nonatomic) SorterType sortType;
 @property (nonatomic) BOOL descendingSort;
